keep syslog identifier in memory
According to the syslog docs, syslog does not store the identifier string
passed when `_openlog` is called- only the pointer. This means that if the
identifier gets garbage-collected within the program, the identifier string
within the logs will be whatever binary nonsense now resides at that memory
location.

Because of this, we can't expose separate calls to open and close the syslog
connection.